Move scyjava.jvm functions into scyjava.config
The scyjava.jvm functions/methods have been
moved to the scyjava.config module.
